chore: organize sorting models
